backend/requirements.txt
fastapi==0.116.1
fuzzywuzzy==0.18.0
langchain==0.3.27
langchain_community==0.3.27
langchain_huggingface==0.3.1
langchain_ollama==0.3.7
langchain_openai==0.3.31
langchain_text_splitters==0.3.9
openai==1.101.0
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.5
uvicorn==0.35.0
sentence-transformers==5.1.0
einops==0.8.1
chromadb==1.0.20